重构数据均分塔FE计算 (#40)

+++ src/main/java/com/fish_dan_/data_energistics/menu/DataDistributionTowerMenu.java
@@ -36,7 +36,7 @@ public class DataDistributionTowerMenu extends AEBaseMenu {
     @GuiSync(731)
     public int maxChannels;
     @GuiSync(732)
-    public int availableFe;
+    public long availableFe;
     @GuiSync(733)
     public int chunkRadius;
     @GuiSync(734)
